+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 ### v0.0.2
 
 * Added `.length` attribute
+* Added `.findByIndx` method
 
 ### v0.0.1