+++ CHANGELOG.md
@@ -1,5 +1,9 @@
 # Change log
 
+### v0.0.3
+
+* Added iterators and other collection processing functions from Underscore.js
+
 ### v0.0.2
 
 * Added `.length` attribute